pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ankibot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.0",
    "python-dotenv",
    "requests",
    "beautifulsoup4",
    "pylint",
    "nltk",
]

[project.scripts]
ankibot-run = "ankibot.__main__:main"
